Criacao Endpoint

+++ src/main/java/br/com/amandareis/CustomerDetail.java
@@ -9,6 +9,7 @@
 package br.com.amandareis;
 
 import java.math.BigInteger;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
@@ -47,7 +48,7 @@ import javax.xml.bind.annotation.XmlType;
 public class CustomerDetail {
 
     @XmlElement(required = true)
-    protected BigInteger id;
+    protected Integer id;
     @XmlElement(required = true)
     protected String name;
     @XmlElement(required = true)
@@ -60,10 +61,10 @@ public class CustomerDetail {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -72,10 +73,10 @@ public class CustomerDetail {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setId(BigInteger value) {
+    public void setId(Integer value) {
         this.id = value;
     }
